hands_on_d/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
cx-Oracle==8.1.0
Django==3.1.6
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-tinymce==3.2.0
future==0.18.2
html5lib==1.1
Pillow==8.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
